+++ packages/tts-editor/CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-01-05
+
 ### Added
 
 - Adds a view that lists the objects in the game that have scripts attached
 - Adds a command to show the current UI of an object as a file
 - Adds a command to locate an object on the table
+- Save and play can be used with the bundled scripts as well
 
 ### Fixed